Add ConfigMigrator
