+++ lib/main.js
@@ -58,5 +58,10 @@ export default {
     const Latex = require('./latex')
     global.latex = new Latex()
     this.disposables.add(global.latex)
+
+    if (!atom.inSpecMode()) {
+      const ConfigMigrator = require('./config-migrator')
+      ConfigMigrator.checkEnvironment()
+    }
   }
 }
